Added analysis that leverages VTK-m for isosurfacing.
Added VTKmContourAnalysis which leverages VTK-m for isosurfacing.
Currently, this analysis works only with the oscillators mini-app
but could be relatively easily adapted to others (minus the part
that generates ghost levels).

VTKmContourAnalysis also leverage DIY to exchange ghost cells.
This is needed because the oscillators miniapp generates cell
data. To remove boundary artifacts, 2 layers of ghost cells are
exchanged - one for point centered isosurfacing, the other for
calculating normals.

Note: There is currently a minor issue with the way VTK-m
generates iso-surfaces. It does not preserve cell data, which
is necessary to mark ghost arrays in the output polydata. This
leads to the visualization of overlapping cells when the output
is displayed. This will be fixed in the future.