autocorelation request mesh with structure
The old ADIOS adaptor ignored the sturture only parameter
in the GetMesh override, the new ADIOS adaptor honors it.
Autocorelation locates the points at whre maximum correlation
occur and hence it needs the mesh structure.

I also replaced a couple of coccurances of cerr with SENSEI_ERROR
and fixed indentation of a couple of small helper classes.