work around FLEXPATH read not understanding constants in global array dims
This works around ADIOS issue #152, FLEXPATH read does not
handles constants in gloabl array dimensions. The work around
is to add some dummy scalar fields to the schema to hold the
constants. If/when ADIOS #152 is fixed we can revert this patch.
Note: do not forget to revert same for array names which was
committed in an earlier patch.