fix version detection regex
it was not working when a tag was checked out. in that
case the match failed because git describe does not
append the SHA at the end. we currently didn't use the
SHA anyway so I changed the regex to match 0 or more
character at the end of the vX.Y.Z