do not SWIG -threads
SWIG's -threads flag tells SWIG to release GIL before entering
all wrapped functions, the idea being that Python can run concurently
with the wrapped C++ code. This is only safe when the wrapped code
aquires the GIL before making calls to the Python API, and the
interpreter has to be initialized for threading. Niether of these
are true by default in ParaView leading to mysterious segv's.
Note, there is a CMake flag that can toggle the correct behavior
at compile time. However, it is disabled by default so we will
not use -threads.