analysis adaptor finalize api
Add a finalize method, this should be called by the bridge
before the analysis is deleted giving the analysis a chance
to clean up and shut down. The motivation is to ease use of
smart pointers with analyses that need to do MPI calls to
clean up and shut down. Without an explicit finalize step
the analysis potentially end up calling MPI after MPI_Finalize
because the smart pointer goes out of scope after MPI_Finalize
is called causing run to abort/crash.