add SliceExtract
this analysis can calculate a slice from a given point and normal or a
set of iso-surfaces defined by a mesh and array spec and list of values.
in both cases the extract is written to disk using the VTKPosthocIO
adaptor.  the analysis is in-transit aware and makes use of the
PlanarSlicePartitioner or the IsoSurfacePartitioner to move only the
necessary blocks