analysis request ghost cells, take 2
Always attempting to add ghost cells is problematic for in transit.
Instead we can use metadata values less than zero to indicate that there are
no ghost zones/nodes but the masking array needs to be added. This for
example an AMR simulation with nested domains but no ghost cells.