Add regression tests for ADISO2 M x N
note about the ADIOS tests: There is a flag to enable adding sender and
receiver block owner arrays to the metadata. These arrays are generated
in the data adaptor and are useful for validating M x N partitioning
behavior. The regression tests color by them. However, reporting them in
the metadata breaks the numbering scheme. Therefor, the flag must be set
manually in the source code and it is off by default. the images
rendered by the tests do not show the partiioning unless the flag is
set. One can however examine the terminal output to verify partitioning
as it is sent to stderr as the test runs.