catalyst bi-directional
Add ability to support bi-direction analysis pipelines to the
CatalystAnalysisAdaptor. Currently, we support result generation for
Python-based Catalyst analysis. User can identify the source that
produces the result mesh and the name to use for that mesh in XML
configuration.

`testOscillatorCatalystBiDirectional` tests this new capability using
the oscillator miniapp. The test simply moves the oscillator centers
using a calculator in the Catalyst Python script.