DataAdaptor remove GetMesh composite data overload
the idea with this overload was to automatically convert to a composite
dataset if the simulation did not return one. The problem was then that
subsequent calls to AddArray etc failed because the simulation did not
expect to be working with a composite dataset. Instead Analysis adaptors
should use SVTKUtils::AsCompositeDataset after all the data has been
fetched from the simulation.