testing fix github actions Python environment macos
CMake found and used a non-default Python install (i.e. that was not in the
runtime environment).  This lead to some runtime errors during regression
tests. This change forces cmake to use the default Python.  We'll also upgrade
from brewed Python 3.9 to the current system default Python install which is
currently Python 3.11.